widl: Add support for top-level conformances.

+++ tools/widl/widl.c
@@ -136,7 +136,7 @@ int line_number = 1;
 
 static FILE *idfile;
 
-size_t pointer_size = 0;
+unsigned int pointer_size = 0;
 syskind_t typelib_kind = sizeof(void*) == 8 ? SYS_WIN64 : SYS_WIN32;
 
 time_t now;
